create separate render/enchance.js tests
